+++ CHANGELOG.md
@@ -36,7 +36,8 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ### Features
 
-* Add grpc messages and cli command to add/remove addresses from metadata scope data access #220
+* Add grpc messages and cli command to add/remove addresses from metadata scope data access [#220](https://github.com/provenance-io/provenance/issues/220)
+* Add a `context` field to the `Session` [#276](https://github.com/provenance-io/provenance/issues/276)
 
 ### Improvements
 
@@ -44,7 +45,8 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ### Bug Fixes
 
-* More mapping fixes related to `WriteP8EContractSpec` and `P8EMemorializeContract` #275
+* More mapping fixes related to `WriteP8EContractSpec` and `P8EMemorializeContract` [#275](https://github.com/provenance-io/provenance/issues/275)
+
 
 ## [v1.2.0](https://github.com/provenance-io/provenance/releases/tag/v1.2.0) - 2021-04-26